ml-services/vision-engine/requirements.txt
fastapi
uvicorn
ultralytics
torch
transformers
Pillow
requests
python-multipart
sentence-transformers
scikit-learn
